// src/types/product.ts
export interface Product {
  id: number;
  brand: string;
  model: string;
  priceINR: number;
  rating: number;
  batteryCapacity: number;
  chargingWattage: number;
  processor: string;
  ram: number;
  storage: number;
  displaySize: number;
  mainCamera: string;
  ultrawideCamera: string;
  telephotoCamera: string;
  selfieCamera: string;
  weight: number;
  imageUrl: string;
}